moved Ruler class to a uicomponent package to better represent what it is supposed to be
